microservices/product-service/pyproject.toml
[project]
name = "product-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=14.0.0",
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "fastapi-pagination>=0.12.34",
    "pydantic-settings>=2.8.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "sqlakeyset>=2.0.1739066250",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
